feat: restrict blog deletion to the owner
Summary:
Check the request for a token, if there isn't one, the request should
fail with a 401 status code. If there is, decode it and check the userId.
If the userId matches with the user (objectId) on the blog, then proceed
ahead and delete the blog, if not, then the user is not authorized the blog,
and the reqeust should fail with a 403