test: fix regression tests
Summary:
introduction of user auth caused some tests to break,
this commit, if applied will fix that, by adding authorization
on the post and delete requests on the /api/blogs endpoint in the
tests. Notice that the test are not optized, since we login to obtain
an access token with each test.